sharedlib/salsanet/requirements.txt
easydict==1.6
joblib==0.10.3
numpy==1.14.4
Pillow>=6.2.0
tensorflow-gpu==1.9.0
scikit-image==0.14.2

